+++ activesupport/CHANGELOG.md
@@ -1,3 +1,18 @@
+*   Fix `RangeError` in `Range#===` and `Range#include?` for endless and beginless ranges
+
+    ```ruby
+    (1..) === (2..)
+    => true
+    (2..) === (1..)
+    => false
+    (..1) === (..2)
+    => false
+    (..2) === (..1)
+    => true
+    ```
+
+    *Alex Mooney*
+
 *   `ActiveSupport::ErrorReporter#report` now assigns a backtrace to unraised exceptions.
 
     Previously reporting an un-raised exception would result in an error report without